% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_api_response}
\alias{check_api_response}
\title{Check API Response Status}
\usage{
check_api_response(resp)
}
\arguments{
\item{resp}{An API response}
}
\description{
Checks the status of an API response
}
