% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_date}
\alias{clean_date}
\title{Clean a provided date}
\usage{
clean_date(date)
}
\arguments{
\item{date}{A date or string purporting to be in \%Y-\%m-\%d format}
}
\description{
Cleans a user provided date, and checks for errors.
}
