% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_mac}
\alias{clean_mac}
\title{Clean a provided MAC number}
\usage{
clean_mac(mac)
}
\arguments{
\item{mac}{The MAC number (also known as GUID/EUI) of the smart meter of interest.}
}
\description{
Cleans a user provided MAC number, and checks for errors.
}
