% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_url}
\alias{create_url}
\title{Create URL for n3rgy API}
\usage{
create_url(query_type, from = NULL, to = NULL)
}
\arguments{
\item{query_type}{One of 'consumption' or 'tariff'.}

\item{from}{(Optional) A date, or string in \%Y-\%m-\%d format representing the earliest date you want data from (n3rgy stores data for 13 months).}

\item{to}{(Optional) A date, or string in \%Y-\%m-\%d format representing the latest date that you want data until.}
}
\description{
Constructs a valid URL to query to the n3rgy API.
}
