% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n3Rgy-package.R
\docType{package}
\name{n3Rgy-package}
\alias{n3Rgy}
\alias{n3Rgy-package}
\title{n3Rgy: Access UK Smartmeter Data From the n3rgy Consumer API}
\description{
A minimal package to get electricity consumption and tariff data from the n3rgy Consumer API.
}
\author{
\strong{Maintainer}: Owen Daniel \email{aaa@bbb.com}

}
\keyword{internal}
