% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{n3rgy_consumption}
\alias{n3rgy_consumption}
\title{Get electricity consumption data from n3rgy API}
\usage{
n3rgy_consumption(mac, from, to)
}
\arguments{
\item{mac}{The MAC number (also known as GUID/EUI) of the smartmeter of interest.}

\item{from}{The earliest date you want data from (n3rgy stores data for 13 months).}

\item{to}{The latest date that you want data until.}
}
\description{
Returns energy consumption data available from n3rgy.
}
