% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{n3rgy_tariff}
\alias{n3rgy_tariff}
\title{Get electricity tariff data from n3rgy API}
\usage{
n3rgy_tariff(mac, from = NULL, to = NULL)
}
\arguments{
\item{mac}{The MAC number (also known as GUID/EUI) of the smart meter of interest.}

\item{from}{(Optional) A date, or string in \%Y-\%m-\%d format representing the earliest date you want data from (n3rgy stores data for 13 months).}

\item{to}{(Optional) A date, or string in \%Y-\%m-\%d format representing the latest date that you want data until.}
}
\description{
Returns electricity tariff data available from n3rgy.
}
